refactor: split doctor package into focused files (#1076)
Extract doctor.go into four focused files:
- types.go: Status, CheckResult, Report, Options + helper methods
- checks_infra.go: system/forge checks (onboarding, git, manifest, adapters, forge)
- checks_deps.go: dependency checks (required tools, required skills)
- checks_config.go: config/capability checks (ontology, adapter registry, retry policies, engine capabilities)

doctor.go is reduced to RunChecks only.
